+++ Godeps/_workspace/src/github.com/docker/docker/pkg/stringid/README.md
@@ -1,0 +1,1 @@
+This package provides helper functions for dealing with string identifiers
