+++ Godeps/_workspace/src/github.com/docker/docker/pkg/stringid/stringid.go
@@ -1,4 +1,4 @@
-package common
+package stringid
 
 import (
 	"crypto/rand"
@@ -35,13 +35,4 @@ func GenerateRandomID() string {
 		}
 		return value
 	}
-}
-
-func RandomString() string {
-	id := make([]byte, 32)
-
-	if _, err := io.ReadFull(rand.Reader, id); err != nil {
-		panic(err) // This shouldn't happen
-	}
-	return hex.EncodeToString(id)
 }
